Added drafts for views functions
